I fixed the ENUMs in the GroupOfCards class, and made Hand a subclass of Deck, and Deck the ONLY subclass of GroupOfCards.
I also fixed up register, since it had troubles with the scanner
not working.

I linked the BlackJack play with the Player play, and put the methods
for the players in the Human class.

I changed the data type of the players arraylist to be Human instead
because I don't want to cast Human every goddamn time I call an element.
This means that the dealer player will be referenced directly.